+++ ChangeLog.md
@@ -5,6 +5,9 @@ Web change log
 
 ## 4.4.0 / 2024-06-10
 
+* Added trace marker (*debug* or *error*) to the server log files when
+  the development console is shown.
+  (@thekid)
 * Merged PR #117: Catch errors and display them in development console.
   (@thekid)
 * Fixed `echo 0;` (or 0.0, or "0") not triggering the development console